backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Full-stack Todo Web Application - Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlmodel>=0.0.22",
    "pydantic>=2.10.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "better-exceptions>=0.3.3",
    "python-dotenv>=1.0.1",
    "asyncpg>=0.30.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.25.0",
    "mypy>=1.13.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
